backend/budgetbox/entries/requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
Brotli==1.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
clerk-backend-api==3.2.1
clerk-sdk-python==0.1.0
colorama==0.4.6
contourpy==1.3.3
cryptography==45.0.6
cycler==0.12.1
distro==1.9.0
Django==5.2.5
djangorestframework==3.16.1
ecdsa==0.19.1
fonttools==4.59.2
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
kiwisolver==1.4.9
matplotlib==3.10.5
multidict==6.6.4
nulltype==2.3.1
numpy==2.3.2
openai==1.102.0
packaging==25.0
pillow==11.3.0
plaid-python==36.0.0
propcache==0.3.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycares==4.10.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
requests==2.32.5
response==0.5.0
rsa==4.9.1
scipy==1.16.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1
